Combine web application factory to integration fixture and use assemblies for db contexts
