+++ Tests/Integration/Fixtures/IntegrationTestFixture.cs
@@ -3,8 +3,12 @@ using Assignment;
 using Auth;
 using Database.Migration;
 using Document;
+using Microsoft.AspNetCore.Hosting;
+using Microsoft.AspNetCore.Mvc.Testing;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 using Notification;
 using Request;
@@ -13,7 +17,7 @@ using Testcontainers.RabbitMq;
 
 namespace Integration.Fixtures;
 
-public class IntegrationTestFixture : IAsyncLifetime
+public class IntegrationTestFixture : WebApplicationFactory<Program>, IAsyncLifetime
 {
     public MsSqlContainer Mssql { get; } = new MsSqlBuilder()
         .WithImage("mcr.microsoft.com/mssql/server:2022-latest")
@@ -106,5 +110,45 @@ public class IntegrationTestFixture : IAsyncLifetime
 
         var service = new MigrationService(databaseMigrator, configuration, migrationLogger);
         await service.MigrateAsync();
+    }
+    protected override void ConfigureWebHost(IWebHostBuilder builder)
+    {
+        Environment.SetEnvironmentVariable("ASPNETCORE_ENVIRONMENT", Environments.Development);
+        builder.UseEnvironment(Environments.Development);
+
+        builder.ConfigureAppConfiguration((context, configBuilder) =>
+        {
+            configBuilder.AddInMemoryCollection(new Dictionary<string, string?>
+            {
+                ["ConnectionStrings:DefaultConnection"] = Mssql.GetConnectionString(),
+                ["ConnectionStrings:Database"] = Mssql.GetConnectionString(),
+                ["RabbitMq:Host"] = RabbitMq.GetConnectionString(),
+                ["RabbitMq:Username"] = "testuser",
+                ["RabbitMq:Password"] = "testpw",
+            });
+        });
+
+        builder.ConfigureServices(ReplaceAllDbContextConnection);
+    }
+
+    private void ReplaceAllDbContextConnection(IServiceCollection services)
+    {
+        var dbContexts = GetAllDbContexts();
+        var replaceMethod = GetType().GetMethod("ReplaceDbContextConnection", BindingFlags.Instance | BindingFlags.NonPublic)!;
+        foreach (var dbContext in dbContexts)
+        {
+            var genericMethod = replaceMethod.MakeGenericMethod(dbContext);
+            genericMethod.Invoke(this, [services]);
+        }
+    }
+
+    private void ReplaceDbContextConnection<T>(IServiceCollection services) where T : DbContext
+    {
+        var descriptor = services.SingleOrDefault(d => d.ServiceType == typeof(DbContextOptions<T>));
+        if (descriptor != null)
+        {
+            services.Remove(descriptor);
+        }
+        services.AddDbContext<T>(options => options.UseSqlServer(Mssql.GetConnectionString()));
     }
 }
